requirements.txt
flask
psycopg2-binary
requests
numpy
opencv-python
gunicorn